refactor: split TransformerPPO agent internals into helpers
